+++ src/InsightBoardReporting.egg-info/SOURCES.txt
@@ -1,5 +1,6 @@
 .gitignore
 .python-version
+LICENSE
 README.md
 main.py
 pyproject.toml
